+++ examples/gunicorn_thrift/ping_app.py
@@ -11,4 +11,5 @@ class Dispatcher(object):
         print("ping pong!")
         return 'pong'
 
+
 app = TProcessor(pingpong.PingService, Dispatcher())