+++ examples/gunicorn_thrift/ping_client.py
@@ -11,5 +11,6 @@ def main():
         pong = c.ping()
         print(pong)
 
+
 if __name__ == '__main__':
     main()