+++ tests/test_socket.py
@@ -70,6 +70,39 @@ class TestTSocket:
         assert client_socket.sock is None
 
 
+class TestTServerSocket:
+    def test_close(self):
+        mock_sock = mock.Mock()
+        server_socket = TServerSocket()
+        server_socket.sock = mock_sock
+        server_socket.close()
+        mock_sock.shutdown.assert_called_once()
+        mock_sock.close.assert_called_once()
+        assert server_socket.sock is None
+
+    def test_close__shutdown_OSError(self):
+        """An OSError on socket shutdown will still close the socket."""
+        mock_sock = mock.Mock()
+        server_socket = TServerSocket()
+        server_socket.sock = mock_sock
+        mock_sock.shutdown.side_effect = OSError
+        server_socket.close()
+        mock_sock.shutdown.assert_called_once()
+        mock_sock.close.assert_called_once()
+        assert server_socket.sock is None
+
+    def test_close__close_OSError(self):
+        """An OSError on socket close will still clear out the socket."""
+        mock_sock = mock.Mock()
+        server_socket = TServerSocket()
+        server_socket.sock = mock_sock
+        mock_sock.close.side_effect = OSError
+        server_socket.close()
+        mock_sock.shutdown.assert_called_once()
+        mock_sock.close.assert_called_once()
+        assert server_socket.sock is None
+
+
 @pytest.mark.skipif(os.getenv('TRAVIS', '') == 'true',
                     reason='Travis CI dose not support IPv6')
 def test_inet6_socket():