+++ thriftpy2/contrib/aio/transport/base.py
@@ -2,17 +2,14 @@
 
 from __future__ import absolute_import
 
-import asyncio
-
 from thriftpy2.transport import TTransportBase, TTransportException
 
 
-@asyncio.coroutine
-def readall(read_fn, sz):
+async def readall(read_fn, sz):
     buff = b''
     have = 0
     while have < sz:
-        chunk = yield from read_fn(sz - have)
+        chunk = await read_fn(sz - have)
         have += len(chunk)
         buff += chunk
 
@@ -31,24 +28,20 @@ class TAsyncTransportBase(TTransportBase):
     def is_open(self):
         raise NotImplementedError
 
-    @asyncio.coroutine
-    def open(self):
+    async def open(self):
         raise NotImplementedError
 
     def close(self):
         raise NotImplementedError
 
-    @asyncio.coroutine
-    def _read(self, sz):
+    async def _read(self, sz):
         raise NotImplementedError
 
-    @asyncio.coroutine
-    def read(self, sz):
-        return (yield from readall(self._read, sz))
+    async def read(self, sz):
+        return await readall(self._read, sz)
 
     def write(self, buf):
         raise NotImplementedError
 
-    @asyncio.coroutine
-    def flush(self):
+    async def flush(self):
         raise NotImplementedError