+++ thriftpy2/protocol/cybin/cybin.pyx
@@ -46,6 +46,7 @@ ctypedef enum TType:
 
 BIN_TYPES = (T_BINARY, T_STRING)
 
+
 class ProtocolError(Exception):
     pass
 
@@ -114,6 +115,9 @@ cdef inline write_list(CyTransportBase buf, object val, spec):
         e_type = spec[0]
         e_spec = spec[1]
 
+    if e_type == T_BINARY:
+        e_type = T_STRING
+
     val_len = len(val)
     write_i08(buf, e_type)
     write_i32(buf, val_len)
@@ -141,6 +145,9 @@ cdef inline write_dict(CyTransportBase buf, object val, spec):
         k_type = key[0]
         k_spec = key[1]
 
+    if k_type == T_BINARY:
+        k_type = T_STRING
+
     value = spec[1]
     if isinstance(value, int):
         v_type = value
@@ -148,6 +155,9 @@ cdef inline write_dict(CyTransportBase buf, object val, spec):
     else:
         v_type = value[0]
         v_spec = value[1]
+
+    if v_type == T_BINARY:
+        v_type = T_STRING
 
     val_len = len(val)
 
@@ -245,7 +255,7 @@ cdef inline c_read_string(CyTransportBase buf, int32_t size):
     py_data = c_read_binary(buf, size)
     try:
         return (<char *>py_data)[:size].decode("utf-8")
-    except:
+    except:  # noqa
         return py_data
 
 
@@ -334,8 +344,11 @@ cdef c_read_val(CyTransportBase buf, TType ttype, spec=None,
                 skip(buf, orig_type)
             return {}
 
-        return {c_read_val(buf, k_type, k_spec, decode_response): c_read_val(buf, v_type, v_spec, decode_response)
-                for _ in range(size)}
+        return {
+            c_read_val(buf, k_type, k_spec, decode_response):
+                c_read_val(buf, v_type, v_spec, decode_response)
+            for _ in range(size)
+        }
 
     elif ttype == T_STRUCT:
         return read_struct(buf, spec(), decode_response)